Refactor Login request flow and QR Scanner
Make the login request view a new screen instead of a modal, which
simplyfies the code and avoids weird border scenarios where a modal
could prevent another modal from showing up.

Login requests are passed directly to the Home/Login screen, instead of
being passed to Home/Home.

The rework on the QRScanner makes its less complex and the interactions
easier to follow. It now exposes just +onRead+ and +onCancel+.

onRead: The scanner will close itself and trigger onRead when it's fully
dismissed. This way the caller doesn't have to worry about waiting for
the modal to be closed.

onCancel: When the user wants to cancel. This event is triggered when
the camera is still open. The caller needs to close the scanner.