Add initial tests for WalletConnect reducer
This adds some simple tests for WalletConnect. These are intended to be
the initial step towards reworking WalletConnect actions and the reducer
itself.

Since this includes no functional changes, it's a low risk
changeset, which should speed up review discussions and leave the
discussion for the refactor isolated from this.