Refactor WalletConnect
This refactor should fix some issues we've been having on the
WalletConnect flow and improve the overall user experience around this
feature. The changes include:

- Avoid timeout after receiving a session request. This prevents the
users from getting a timeout while looking at the confirmation window.
- Fixed feedback when waiting for a session request, which will help
understand the application status when the WalletConnect connection is
being initialized.
- Sessions are now persisted with redux-persist (using
walletConnectSessions reducer).
- Old sessions will be imported to the new reducer.
- Approving a request for a url will disconnect old clients for that
url. This is so we don't have a way of terminating old sessions, a more
sensible approach might be possible.
- Correctly handle multiple requests on the Requests List view.