Allow callbacks to be passed to Toast messages
Toast messages will receive an +onPress+ callback which, if passed, will
be called when the user presses on the Toast message text. This allows
making the Toast more useful.

The wallet backup message will now use the Toast callback to redirect
the user to the page where the backup functionality lives. The message
text itself was also updated to reflect the new location of the backup
feature.